Get the account transaction history
I don't like that the same ViewModel is being used for both Fragments; I also don't like that the VM's have the LiveData while the repo doesn't. This works for now, because all the data is local, but a real app would have to fetch from the network, so the LiveData ought to be pushed down a level.